Track double > float

+++ track.cpp
@@ -20,7 +20,7 @@ Track::Track(QWidget *parent) : QWidget(parent)
             this, SLOT(on_zoomChanged(QCPRange)));
 }
 
-void Track::plot(std::vector<double> vec)
+void Track::plot(std::vector<float> vec)
 {
     plotter->plot(vec);
 }
